+++ doc/changelog.md
@@ -1,3 +1,10 @@
+## 0.8.1
+#### element.js Changes
+ - Added Error Checking for element dependencies 
+ - Added a new flag called `strictStyles` to control a new feature
+#### documentation Changes
+ - updated the documentation for [element.js](element.js.md)
+ - added new file called [Making Elements](Making%20Elements.md) just for explaining the process of making elements
 # 0.8
 
 #### tileWin.js Changes